rename miscLidar.py and move global_settings.py to utils
